requirements.txt
python-telegram-bot==20.7
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
Flask